prevents dashboard counts from reloading with line chart filters change
